requirements.txt
# Put here all your python libs dependencies before vtenv.sh
django
djangorestframework
djangorestframework-jwt
django-extensions
django-debug-toolbar >= 0.8.5
django-autoslug
# https://code.djangoproject.com/ticket/16250
#psycopg2==2.4.1
django-environ
django-rest-swagger
